Fix the dv store dtype I broke in 503842a
503842a pinned the dAv kernel's `dv` output to f32 while letting the cotangent
stay bf16, but left the store as

    dv_ref[:] = b_dv.astype(do_ref.dtype)

Those two agreed until the cotangent stopped being f32. On hardware it fails at
Pallas tracing:

    ValueError: Invalid dtype for `swap`. Ref dtype: float32.
                Value dtype: bfloat16.

The store now uses `dv_ref.dtype`, which is what it should have said. Tying a
store to an *input* ref's dtype is the bug; the output's own dtype is the only
correct source.

Why no CPU gate caught it, which matters more than the one-line fix.
`check_tpu_lowering` and `check_remat_gradients` both run float32 only, and in
float32 `do_ref.dtype == dv_ref.dtype`, so the two agreed and nothing
complained. Production runs bf16. **Every CPU gate in this repo is blind to
bf16-only bugs.**

I tried to close it by adding bf16 cases to `check_tpu_lowering` and could not:
the bf16 path calls `jax._src.tpu_info.get_tpu_info` from inside lowering, and
patching the `pltpu` alias the file already fakes does not reach it. Rather
than leave a gate that fails six of twelve cases for an unrelated reason, the
limitation is now written at the top of the file along with the pattern to grep
for, and the run prints it.

Verified on a CPU host: gradients unchanged in both gate modes, all six
configurations still lower. The failure this fixes only appears in bf16, so
neither gate can confirm the fix either. That is the point of the note.

+++ tokamax/_src/ops/experimental/kda/pallas_mosaic_tpu_bwd_kernel.py
@@ -1456,7 +1456,10 @@ def _chunk_kda_bwd_dAv_kernel(
   b_dA = jnp.where(m_causal[None, :, :], b_dA * scale, 0.0)
 
   dA_ref[:] = b_dA
-  dv_ref[:] = b_dv.astype(do_ref.dtype)
+  # `dv_ref.dtype`, not `do_ref.dtype`. The output is pinned to f32 because it
+  # is a partial gradient; the cotangent is now bf16. Tying the store to the
+  # input dtype makes them disagree the moment those two differ.
+  dv_ref[:] = b_dv.astype(dv_ref.dtype)
 
 
 @functools.partial(
